Initial code to run manticore
In the `security` directory, there is a `manticore` directory that
includes some scripts. They are invoked by the new command
`npm run-script manticore`.

`anchor.py` includes a comment that should make temporary flattening the
contracts unnecessary.
The change to `CircularBufferUint` should also not be necessary. Check
if it would work with remix and truffle and raise an issue with
manticore if it doesn't work with manticore.

References #471